Move interfaces to shared models
