+++ core-web/libs/portlets/dot-ema/src/lib/edit-ema-editor/edit-ema-editor.component.ts
@@ -32,15 +32,13 @@ import { EditEmaPersonaSelectorComponent } from './components/edit-ema-persona-s
 import { EditEmaToolbarComponent } from './components/edit-ema-toolbar/edit-ema-toolbar.component';
 
 import {
-    ActionPayload,
     EmaPageDropzoneComponent,
-    PlacePayload,
     Row
 } from '../components/ema-page-dropzone/ema-page-dropzone.component';
 import { EditEmaStore } from '../dot-ema-shell/store/dot-ema.store';
 import { DEFAULT_LANGUAGE_ID, DEFAULT_PERSONA, DEFAULT_URL, HOST, WINDOW } from '../shared/consts';
 import { CUSTOMER_ACTIONS, NG_CUSTOM_EVENTS, NOTIFY_CUSTOMER } from '../shared/enums';
-import { SetUrlPayload } from '../shared/models';
+import { ActionPayload, PlacePayload, SetUrlPayload } from '../shared/models';
 import {
     deleteContentletFromContainer,
     insertContentletInContainer,
